Changed To Unitless Line Height
- Line Height changed from percentage to unitless as this should result in more predictable theming.
- Line Height now defaults to being unset.
Attempts to fix issue: https://github.com/btpf/Alexandria/issues/50

+++ src-tauri/src/main.rs
@@ -472,7 +472,7 @@ struct themePayload {
     #[serde(default)]
     wordSpacing: i64,
     #[serde(default)]
-    lineHeight: i64,
+    lineHeight: f64,
     #[serde(default)]
     renderMode: String,
     #[serde(default)]
